// apps/web-admin/src/api/model/createUserDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * 采摘通 API
 * CaiZhaiTong Full Lifecycle Management System API
 * OpenAPI spec version: 1.0
 */
import type { CreateUserDtoRoleKey } from './createUserDtoRoleKey';

export interface CreateUserDto {
  /** 真实姓名 */
  name: string;
  /** 身份证号 */
  idCard: string;
  /** 手机号 */
  phone: string;
  /** 角色 */
  roleKey: CreateUserDtoRoleKey;
  faceImgUrl?: string;
  /** 区域管理员需填区域码 */
  regionCode?: number;
}