// apps/web-admin/src/api/model/createUserDtoRoleKey.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * 采摘通 API
 * CaiZhaiTong Full Lifecycle Management System API
 * OpenAPI spec version: 1.0
 */

/**
 * 角色
 */
export type CreateUserDtoRoleKey = typeof CreateUserDtoRoleKey[keyof typeof CreateUserDtoRoleKey];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateUserDtoRoleKey = {
  super_admin: 'super_admin',
  region_admin: 'region_admin',
  base_manager: 'base_manager',
  field_manager: 'field_manager',
  worker: 'worker',
} as const;